:sparkles: feat: finishing CRUD Publisher

+++ src/controller/publisherController.ts
@@ -4,13 +4,16 @@ import AppDataSource from '../database/connection';
 import Publisher from '../models/publisherModel';
 import PublisherView from '../views/publisherView';
 import { publisherSchema } from '../schemas/validators';
+import BookController from './bookController';
 
 class PublisherController {
   static async getAllPublishers(req: Request, res: Response) {
     try {
       const publisherRepository = AppDataSource.getRepository(Publisher);
 
-      const publishers = await publisherRepository.find();
+      const publishers = await publisherRepository.find({ // O método find quando não tem condição, lista tudo
+        relations: ['books']
+      });
 
       return res.json(PublisherView.renderMany(publishers));
 
@@ -25,7 +28,10 @@ class PublisherController {
     try {
       const publisherRepository = AppDataSource.getRepository(Publisher);
 
-      const publisher = await publisherRepository.findOneBy({ id: parseInt(id) });
+      const publisher = await publisherRepository.findOne({
+        relations: ['books'],
+        where: { id: parseInt(id) }
+      });
 
       if (!publisher) {
         return res.status(404).json({ message: 'Editora não encontrad' });
@@ -38,7 +44,7 @@ class PublisherController {
   };
 
   static async createPublisher(req: Request, res: Response) {
-    const { name, books } = req.body;
+    const { name } = req.body;
 
     try {
 
@@ -57,37 +63,32 @@ class PublisherController {
     }
   };
 
-  // static async updatePublisher(req: Request, res: Response) {
-  //   const { id } = req.params;
-  //   const { title, author, isbn, yearOfPublication } = req.body;
+  static async updatePublisher(req: Request, res: Response) {
+    const { id } = req.params;
+    const { name } = req.body;
 
-  //   try {
-  //     const publisherRepository = AppDataSource.getRepository(Publisher);
+    try {
+      const publisherRepository = AppDataSource.getRepository(Publisher);
 
-  //     const publisher = await publisherRepository.findOneBy({ id: parseInt(id) });
+      const publisher = await publisherRepository.findOneBy({ id: parseInt(id) });
 
-  //     if (!publisher) {
-  //       return res.status(404).json({ message: 'Livro não encontrado' });
-  //     }
+      if (!publisher) {
+        return res.status(404).json({ message: 'Editora não encontrada' });
+      }
 
-  //     const data = { title, author, isbn, yearOfPublication };
+      const data = { name };
 
-  //     await schema.validate(data, {
-  //       abortEarly: false,
-  //     });
+      await publisherSchema.validate(data, { abortEarly: false });
 
-  //     book.title = title;
-  //     book.author = author;
-  //     book.isbn = isbn;
-  //     book.yearOfPublication = yearOfPublication;
+      publisher.name = name;
 
-  //     await publisherRepository.save(publisher);
+      await publisherRepository.save(publisher);
 
-  //     return res.json(PublisherView.render(publisher));
-  //   } catch (err: any) {
-  //     return res.status(400).json({ message: err.message });
-  //   }
-  // };
+      return res.json(PublisherView.render(publisher));
+    } catch (err: any) {
+      return res.status(400).json({ message: err.message });
+    }
+  };
 
   static async deletePublisher(req: Request, res: Response) {
     const { id } = req.params;
@@ -101,11 +102,29 @@ class PublisherController {
         return res.status(404).json({ message: 'Livro não encontrado' });
       }
 
+      const publisherHasBooks = await BookController.publisherHasBooks(publisher);
+      if (publisherHasBooks) {
+        return res.status(405).json({ message: 'Editora possui livros atrelados' });
+      }
+
       await publisherRepository.remove(publisher);
 
       return res.status(204).json();
     } catch (err: any) {
       return res.status(500).json({ message: err.message });
+    }
+  };
+
+  static async checkIfPublisherExists(id: number): Promise<boolean> {
+    try {
+      const publisherRepository = AppDataSource.getRepository(Publisher);
+
+      const publisher = await publisherRepository.findOneBy({ id: id });
+
+      return publisher ? true : false;
+
+    } catch (err: any) {
+      throw err;
     }
   };
 }
